Window.rmhmodel {spatstat.random}	R Documentation

Extract Window of Spatial Object

Description

Given a spatial object (such as a point pattern or pixel image) in two dimensions, these functions extract the window in which the object is defined.

Usage

 ## S3 method for class 'rmhmodel'
Window(X, ...)

Arguments

X
A spatial object.

...
Ignored.

Details

These are methods for the generic function Window which extract the spatial window in which the object X is defined.

Value

An object of class "owin" (see owin.object) specifying an observation window.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

Window, Window.ppp, Window.psp.

owin.object

Examples

   A <- rmhmodel(cif='poisson', par=list(beta=10), w=square(2))
   Window(A)

[Package spatstat.random version 3.3-2 Index]