clusterkernel {spatstat.random}	R Documentation

Extract Cluster Offspring Kernel

Description

Given a cluster point process model, this command returns the probability density of the cluster offspring.

Usage

clusterkernel(model, ...)
## S3 method for class 'character'
clusterkernel(model, ...)

Arguments

model
Cluster model. Either a fitted cluster or Cox model (object of class "kppm"), or a character string specifying the type of cluster model.

...
Parameter values for the model, when model is a character string.

Details

Given a specification of a cluster point process model, this command returns a function(x,y) giving the two-dimensional probability density of the cluster offspring points assuming a cluster parent located at the origin.

The function clusterkernel is generic, with methods for class "character" (described here) and "kppm" (described in clusterkernel.kppm).

Value

A function in the R language with arguments x,y,....

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

clusterkernel.kppm), clusterfield, kppm.

Examples

  f <- clusterkernel("Thomas", kappa=10, scale=0.5)
  f(0.1, 0.2)

[Package spatstat.random version 3.3-2 Index]