default.rmhcontrol {spatstat.random}	R Documentation

Set Default Control Parameters for Metropolis-Hastings Algorithm.

Description

For a Gibbs point process model (either a fitted model, or a model specified by its parameters), this command sets appropriate default values of the parameters controlling the iterative behaviour of the Metropolis-Hastings algorithm.

Usage

   default.rmhcontrol(model, w=NULL)

Arguments

model
A fitted point process model (object of class "ppm") or a description of a Gibbs point process model (object of class "rmhmodel").

w
Optional. Window for the resulting simulated patterns.

Details

This function sets the values of the parameters controlling the iterative behaviour of the Metropolis-Hastings simulation algorithm. It uses default values that would be appropriate for the fitted point process model model.

The expansion parameter expand is set to default.expand(model, w).

All other parameters revert to their defaults given in rmhcontrol.default.

See rmhcontrol for the full list of control parameters. To override default parameters, use update.rmhcontrol.

Value

An object of class "rmhcontrol". See rmhcontrol.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

See Also

rmhcontrol, update.rmhcontrol, ppm, default.expand

Examples

if(require(spatstat.model)) {
  fit <- ppm(cells, ~1, Strauss(0.1))
  default.rmhcontrol(fit)
  default.rmhcontrol(fit, w=square(2))
}
   m <- rmhmodel(cif='strauss',
                 par=list(beta=100, gamma=0.5, r=0.1),
                 w=unit.square())
  default.rmhcontrol(m)
  default.rmhcontrol(m, w=square(2))

[Package spatstat.random version 3.3-2 Index]