dpakes {spatstat.random}	R Documentation

Pakes distribution

Description

Probability density, cumulative distribution function, quantile function, and random generation for the Pakes distribution.

Usage

dpakes(x, zeta)
ppakes(q, zeta)
qpakes(p, zeta)
rpakes(n, zeta)

Arguments

x,q
Numeric vector of quantiles.

p
Numeric vector of probabilities

n
Number of observations.

zeta
Mean of distribution. A single, non-negative, numeric value.

Details

These functions concern the probability distribution of the random variable

X = sum_[n=1]^[infty] prod_[j=1]^[n] U[j]^(1/zeta)

where U[1], U[2], ... are independent random variables uniformly distributed on [0,1] and zeta is a parameter.

This distribution arises in many contexts. For example, for a homogeneous Poisson point process in two-dimensional space with intensity lambda, the standard Gaussian kernel estimator of intensity with bandwidth σ, evaluated at any fixed location u, has the same distribution as (lambda/zeta) * X where zeta = 2 * pi * lambda * sigma^2.

Following the usual convention, dpakes computes the probability density, ppakes the cumulative distribution function, and qpakes the quantile function, and rpakes generates random variates with this distribution.

The computation is based on a recursive integral equation for the cumulative distribution function, due to Professor Tony Pakes, presented in Baddeley, Moller and Pakes (2008). The solution uses the fact that the random variable satisfies the distributional equivalence

X == U^(1/zeta) * (1+X)

where U is uniformly distributed on [0,1] and independent of X.

Value

A numeric vector.

Author(s)

Adrian Baddeley.

References

Baddeley, A., Moller, J. and Pakes, A.G. (2008) Properties of residuals for spatial point processes, Annals of the Institute of Statistical Mathematics 60, 627–649.

Examples

 curve(dpakes(x, 1.5), to=4)
 rpakes(3, 1.5)

[Package spatstat.random version 3.3-2 Index]