expand.owin {spatstat.random}	R Documentation

Apply Expansion Rule

Description

Applies an expansion rule to a window.

Usage

 expand.owin(W, ...)

Arguments

W
A window.

...
Arguments passed to rmhexpand to determine an expansion rule.

Details

The argument W should be a window (an object of class "owin").

This command applies the expansion rule specified by the arguments ... to the window W, yielding another window.

The arguments ... are passed to rmhexpand to determine the expansion rule.

For other transformations of the scale, location and orientation of a window, see shift, affine and rotate.

Value

A window (object of class "owin").

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

See Also

rmhexpand about expansion rules.

shift, rotate, affine for other types of manipulation.

Examples

   expand.owin(square(1), 9)
   expand.owin(square(1), distance=0.5)
   expand.owin(letterR, length=2)
   expand.owin(letterR, distance=0.1)

[Package spatstat.random version 3.3-2 Index]