rGaussPoisson {spatstat.random}	R Documentation

Simulate Gauss-Poisson Process

Description

Generate a random point pattern, a simulated realisation of the Gauss-Poisson Process.

Usage

 rGaussPoisson(kappa, r, p2, win = owin(c(0,1),c(0,1)),
               ..., nsim=1, drop=TRUE)

Arguments

kappa
Intensity of the Poisson process of cluster centres. A single positive number, a function, or a pixel image.

r
Diameter of each cluster that consists of exactly 2 points.

p2
Probability that a cluster contains exactly 2 points.

win
Window in which to simulate the pattern. An object of class "owin" or something acceptable to as.owin.

...
Ignored.

nsim
Number of simulated realisations to be generated.

drop
Logical. If nsim=1 and drop=TRUE (the default), the result will be a point pattern, rather than a list containing a point pattern.

Details

This algorithm generates a realisation of the Gauss-Poisson point process inside the window win. The process is constructed by first generating a Poisson point process of parent points with intensity kappa. Then each parent point is either retained (with probability 1 - p2) or replaced by a pair of points at a fixed distance r apart (with probability p2). In the case of clusters of 2 points, the line joining the two points has uniform random orientation.

In this implementation, parent points are not restricted to lie in the window; the parent process is effectively the uniform Poisson process on the infinite plane.

Value

A point pattern (an object of class "ppp") if nsim=1, or a list of point patterns if nsim > 1.

Additionally, some intermediate results of the simulation are returned as attributes of the point pattern. See rNeymanScott.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

See Also

rpoispp, rThomas, rMatClust, rNeymanScott

Examples

 pp <- rGaussPoisson(30, 0.07, 0.5)

[Package spatstat.random version 3.3-2 Index]