rMaternII {spatstat.random}	R Documentation

Simulate Matern Model II

Description

Generate a random point pattern, a simulated realisation of the Matern Model II inhibition process.

Usage

 rMaternII(kappa, r, win = owin(c(0,1),c(0,1)), stationary=TRUE, ...,
           nsim=1, drop=TRUE)

Arguments

kappa
Intensity of the Poisson process of proposal points. A single positive number.

r
Inhibition distance.

win
Window in which to simulate the pattern. An object of class "owin" or something acceptable to as.owin. Alternatively a higher-dimensional box of class "box3" or "boxx".

stationary
Logical. Whether to start with a stationary process of proposal points (stationary=TRUE) or to generate the proposal points only inside the window (stationary=FALSE).

...
Ignored.

nsim
Number of simulated realisations to be generated.

drop
Logical. If nsim=1 and drop=TRUE (the default), the result will be a point pattern, rather than a list containing a point pattern.

Details

This algorithm generates one or more realisations of Matern's Model II inhibition process inside the window win.

The process is constructed by first generating a uniform Poisson point process of “proposal” points with intensity kappa. If stationary = TRUE (the default), the proposal points are generated in a window larger than win that effectively means the proposals are stationary. If stationary=FALSE then the proposal points are only generated inside the window win.

Then each proposal point is marked by an “arrival time”, a number uniformly distributed in [0,1] independently of other variables.

A proposal point is deleted if it lies within r units' distance of another proposal point that has an earlier arrival time. Otherwise it is retained. The retained points constitute Matern's Model II.

The difference between Matern's Model I and II is the italicised statement above. Model II has a higher intensity for the same parameter values.

Value

A point pattern if nsim=1, or a list of point patterns if nsim > 1. Each point pattern is normally an object of class "ppp", but may be of class "pp3" or "ppx" depending on the window.

Author(s)

Ute Hahn, Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

rpoispp, rMatClust, rMaternI

Examples

 X <- rMaternII(20, 0.05)
 Y <- rMaternII(20, 0.05, stationary=FALSE)

[Package spatstat.random version 3.3-2 Index]