rMosaicSet {spatstat.random}	R Documentation

Mosaic Random Set

Description

Generate a random set by taking a random selection of tiles of a given tessellation.

Usage

rMosaicSet(X, p=0.5)

Arguments

X
A tessellation (object of class "tess").

p
Probability of including a given tile. A number strictly between 0 and 1.

Details

Given a tessellation X, this function randomly selects some of the tiles of X, including each tile with probability p independently of the other tiles. The selected tiles are then combined to form a set in the plane.

One application of this is Switzer's (1965) example of a random set which has a Markov property. It is constructed by generating X according to a Poisson line tessellation (see rpoislinetess).

Value

A window (object of class "owin").

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

References

Switzer, P. A random set process in the plane with a Markovian property. Annals of Mathematical Statistics 36 (1965) 1859–1863.

See Also

rpoislinetess, rMosaicField

Examples

    if(interactive()) {
     lambda <- 3
     n <- 30
   } else {
     lambda <- 1
     n <- 5
   }
  # Switzer's random set
   X <- rpoislinetess(lambda)
   plot(rMosaicSet(X, 0.5), col="green", border=NA)

   # another example
   Y <- dirichlet(runifpoint(n))
   plot(rMosaicSet(Y, 0.4))

[Package spatstat.random version 3.3-2 Index]