ragsAreaInter {spatstat.random}	R Documentation

Alternating Gibbs Sampler for Area-Interaction Process

Description

Generate a realisation of the area-interaction process using the alternating Gibbs sampler. Applies only when the interaction parameter eta is greater than 1.

Usage

     ragsAreaInter(beta, eta, r, ...,
                   win = NULL, bmax = NULL, periodic = FALSE, ncycles = 100)

Arguments

beta
First order trend. A number, a pixel image (object of class "im"), or a function(x,y).

eta
Interaction parameter (canonical form) as described in the help for AreaInter. A number greater than 1.

r
Disc radius in the model. A number greater than 1.

...
Additional arguments for beta if it is a function.

win
Simulation window. An object of class "owin". (Ignored if beta is a pixel image.)

bmax
Optional. The maximum possible value of beta, or a number larger than this.

periodic
Logical value indicating whether to treat opposite sides of the simulation window as being the same, so that points close to one side may interact with points close to the opposite side. Feasible only when the window is a rectangle.

ncycles
Number of cycles of the alternating Gibbs sampler to be performed.

Details

This function generates a simulated realisation of the area-interaction process (see AreaInter) using the alternating Gibbs sampler (see rags).

It exploits a mathematical relationship between the (unmarked) area-interaction process and the two-type hard core process (Baddeley and Van Lieshout, 1995; Widom and Rowlinson, 1970). This relationship only holds when the interaction parameter eta is greater than 1 so that the area-interaction process is clustered.

The parameters beta,eta are the canonical parameters described in the help for AreaInter. The first order trend beta may be a constant, a function, or a pixel image.

The simulation window is determined by beta if it is a pixel image, and otherwise by the argument win (the default is the unit square).

Value

A point pattern (object of class "ppp").

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au.

References

Baddeley, A.J. and Van Lieshout, M.N.M. (1995). Area-interaction point processes. Annals of the Institute of Statistical Mathematics 47 (1995) 601–619.

Widom, B. and Rowlinson, J.S. (1970). New model for the study of liquid-vapor phase transitions. The Journal of Chemical Physics 52 (1970) 1670–1684.

See Also

rags, ragsMultiHard

AreaInter

Examples

   plot(ragsAreaInter(100, 2, 0.07, ncycles=15))

[Package spatstat.random version 3.3-2 Index]