reach {spatstat.random}	R Documentation

Interaction Distance of a Point Process Model

Description

Computes the interaction distance of a point process model.

Usage

  reach(x, ...)

  ## S3 method for class 'rmhmodel'
reach(x, ...)

Arguments

x
Either a fitted point process model (object of class "ppm"), an interpoint interaction (object of class "interact"), a fitted interpoint interaction (object of class "fii") or a point process model for simulation (object of class "rmhmodel").

...
Other arguments are ignored.

Details

The function reach computes the ‘interaction distance’ or ‘interaction range’ of a point process model.

The definition of the interaction distance depends on the type of point process model. This help page explains the interaction distance for a Gibbs point process. For other kinds of models, see reach.kppm and reach.dppm.

For a Gibbs point process model, the interaction distance is the shortest distance D such that any two points in the process which are separated by a distance greater than D do not interact with each other.

For example, the interaction range of a Strauss process (see Strauss or rStrauss) with parameters beta,gamma,r is equal to r, unless gamma=1 in which case the model is Poisson and the interaction range is 0. The interaction range of a Poisson process is zero. The interaction range of the Ord threshold process (see OrdThresh) is infinite, since two points may interact at any distance apart.

The function reach is generic, with methods for the case where x is

a fitted point process model (object of class "ppm", usually obtained from the model-fitting function ppm);

an interpoint interaction structure (object of class "interact")

a fitted interpoint interaction (object of class "fii")

a point process model for simulation (object of class "rmhmodel"), usually obtained from rmhmodel.

Value

The interaction distance, or NA if this cannot be computed from the information given.

Other types of models

Methods for reach are also defined for point process models of class "kppm" and "dppm". Their technical definition is different from this one. See reach.kppm and reach.dppm.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

See Also

reach.ppm

rmhmodel

See reach.kppm and reach.dppm for other types of point process models.

Examples

    reach(rmhmodel(cif='poisson', par=list(beta=100)))
    # returns 0

    reach(rmhmodel(cif='strauss', par=list(beta=100, gamma=0.1, r=7)))
    # returns 7

    reach(rmhmodel(cif='sftcr', par=list(beta=100, sigma=1, kappa=0.7)))
    # returns Inf
    
    reach(rmhmodel(cif='multihard',
                   par=list(beta=c(10,10), hradii=matrix(c(1,3,3,1),2,2))))
    # returns 3

[Package spatstat.random version 3.3-2 Index]