rmhstart {spatstat.random}	R Documentation

Determine Initial State for Metropolis-Hastings Simulation.

Description

Builds a description of the initial state for the Metropolis-Hastings algorithm.

Usage

   rmhstart(start, ...)
   ## Default S3 method:
rmhstart(start=NULL, ..., n.start=NULL, x.start=NULL)

Arguments

start
An existing description of the initial state in some format. Incompatible with the arguments listed below.

...
There should be no other arguments.

n.start
Number of initial points (to be randomly generated). Incompatible with x.start.

x.start
Initial point pattern configuration. Incompatible with n.start.

Details

Simulated realisations of many point process models can be generated using the Metropolis-Hastings algorithm implemented in rmh.

This function rmhstart creates a full description of the initial state of the Metropolis-Hastings algorithm, including possibly the initial state of the random number generator, for use in a subsequent call to rmh. It also checks that the initial state is valid.

The initial state should be specified either by the first argument start or by the other arguments n.start, x.start etc.

If start is a list, then it should have components named n.start or x.start, with the same interpretation as described below.

The arguments are:

n.start

The number of “initial” points to be randomly (uniformly) generated in the simulation window w. Incompatible with x.start.

For a multitype point process, n.start may be a vector (of length equal to the number of types) giving the number of points of each type to be generated.

If expansion of the simulation window is selected (see the argument expand to rmhcontrol), then the actual number of starting points in the simulation will be n.start multiplied by the expansion factor (ratio of the areas of the expanded window and original window).

For faster convergence of the Metropolis-Hastings algorithm, the value of n.start should be roughly equal to (an educated guess at) the expected number of points for the point process inside the window.

x.start

Initial point pattern configuration. Incompatible with n.start.

x.start may be a point pattern (an object of class ppp), or an object which can be coerced to this class by as.ppp, or a dataset containing vectors x and y.

If x.start is specified, then expansion of the simulation window (the argument expand of rmhcontrol) is not permitted.

The parameters n.start and x.start are incompatible.

Value

An object of class "rmhstart", which is essentially a list of parameters describing the initial point pattern and (optionally) the initial state of the random number generator.

There is a print method for this class, which prints a sensible description of the initial state.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

See Also

rmh, rmhcontrol, rmhmodel

Examples

   # 30 random points
   a <- rmhstart(n.start=30)
   a

   # a particular point pattern
   b <- rmhstart(x.start=cells)

[Package spatstat.random version 3.3-2 Index]