rpoislinetess {spatstat.random}	R Documentation

Poisson Line Tessellation

Description

Generate a tessellation delineated by the lines of the Poisson line process

Usage

rpoislinetess(lambda, win = owin())

Arguments

lambda
Intensity of the Poisson line process. A positive number.

win
Window in which to simulate the pattern. An object of class "owin" or something acceptable to as.owin. Currently, the window must be a rectangle.

Details

This algorithm generates a realisation of the uniform Poisson line process, and divides the window win into tiles separated by these lines.

The argument lambda must be a positive number. It controls the intensity of the process. The expected number of lines intersecting a convex region of the plane is equal to lambda times the perimeter length of the region. The expected total length of the lines crossing a region of the plane is equal to lambda * pi times the area of the region.

Value

A tessellation (object of class "tess").

Also has an attribute "lines" containing the realisation of the Poisson line process, as an object of class "infline".

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net.

See Also

rpoisline to generate the lines only.

Examples

 X <- rpoislinetess(3)
 plot(as.im(X), main="rpoislinetess(3)")
 plot(X, add=TRUE)

[Package spatstat.random version 3.3-2 Index]