rpoisppx {spatstat.random}	R Documentation

Generate Poisson Point Pattern in Any Dimensions

Description

Generate a random multi-dimensional point pattern using the homogeneous Poisson process.

Usage

rpoisppx(lambda, domain, nsim=1, drop=TRUE)

Arguments

lambda
Intensity of the Poisson process. A single positive number.

domain
Multi-dimensional box in which the process should be generated. An object of class "boxx".

nsim
Number of simulated realisations to be generated.

drop
Logical. If nsim=1 and drop=TRUE (the default), the result will be a point pattern, rather than a list containing a single point pattern.

Details

This function generates a realisation of the homogeneous Poisson process in multi dimensions, with intensity lambda (points per unit volume).

The realisation is generated inside the multi-dimensional region domain which currently must be a rectangular box (object of class "boxx").

Value

If nsim = 1 and drop=TRUE, a point pattern (an object of class "ppx"). If nsim > 1 or drop=FALSE, a list of such point patterns.

Note

The intensity lambda is the expected number of points per unit volume.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net.

See Also

runifpointx, ppx, boxx

Examples

   w <- boxx(x=c(0,1), y=c(0,1), z=c(0,1), t=c(0,3))
   X <- rpoisppx(10, w)

[Package spatstat.random version 3.3-2 Index]