rshift {spatstat.random}	R Documentation

Random Shift

Description

Randomly shifts the points of a point pattern or line segment pattern. Generic.

Usage

   rshift(X, ...)

Arguments

X
Pattern to be subjected to a random shift. A point pattern (class "ppp"), a line segment pattern (class "psp") or an object of class "splitppp".

...
Arguments controlling the generation of the random shift vector, or specifying which parts of the pattern will be shifted.

Details

This operation applies a random shift (vector displacement) to the points in a point pattern, or to the segments in a line segment pattern.

The argument X may be

a point pattern (an object of class "ppp")

a line segment pattern (an object of class "psp")

an object of class "splitppp" (basically a list of point patterns, obtained from split.ppp).

The function rshift is generic, with methods for the three classes "ppp", "psp" and "splitppp".

See the help pages for these methods, rshift.ppp, rshift.psp and rshift.splitppp, for further information.

Value

An object of the same type as X.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net.

See Also

rshift.ppp, rshift.psp, rshift.splitppp

[Package spatstat.random version 3.3-2 Index]