rstrat {spatstat.random}	R Documentation

Simulate Stratified Random Point Pattern

Description

Generates a “stratified random” pattern of points in a window, by dividing the window into rectangular tiles and placing k random points independently in each tile.

Usage

 rstrat(win=square(1), nx, ny=nx, k = 1, nsim=1, drop=TRUE)

Arguments

win
A window. An object of class owin, or data in any format acceptable to as.owin().

nx
Number of tiles in each column.

ny
Number of tiles in each row.

k
Number of random points to generate in each tile.

nsim
Number of simulated realisations to be generated.

drop
Logical. If nsim=1 and drop=TRUE (the default), the result will be a point pattern, rather than a list containing a point pattern.

Details

This function generates a random pattern of points in a “stratified random” sampling design. It can be useful for generating random spatial sampling points.

The bounding rectangle of win is divided into a regular nx * ny grid of rectangular tiles. In each tile, k random points are generated independently with a uniform distribution in that tile.

Some of these grid points may lie outside the window win: if they do, they are deleted.

The result is a point pattern inside the window win.

This function is useful in creating dummy points for quadrature schemes (see quadscheme) as well as in simulating random point patterns.

Value

A point pattern (an object of class "ppp") if nsim=1, or a list of point patterns if nsim > 1.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net.

See Also

rsyst, runifpoint, quadscheme

Examples

  X <- rstrat(nx=10)
  plot(X)

  # polygonal boundary
  X <- rstrat(letterR, 5, 10, k=3)
  plot(X)

[Package spatstat.random version 3.3-2 Index]