runifpointx {spatstat.random}	R Documentation

Generate N Uniform Random Points in Any Dimensions

Description

Generate a random point pattern containing n independent, uniform random points in any number of spatial dimensions.

Usage

runifpointx(n, domain, nsim=1, drop=TRUE)

Arguments

n
Number of points to be generated.

domain
Multi-dimensional box in which the process should be generated. An object of class "boxx".

nsim
Number of simulated realisations to be generated.

drop
Logical. If nsim=1 and drop=TRUE (the default), the result will be a point pattern, rather than a list containing a single point pattern.

Details

This function generates a pattern of n independent random points, uniformly distributed in the multi-dimensional box domain.

Value

If nsim = 1 and drop=TRUE, a point pattern (an object of class "ppx"). If nsim > 1 or drop=FALSE, a list of such point patterns.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net.

See Also

rpoisppx, ppx, boxx

Examples

   w <- boxx(x=c(0,1), y=c(0,1), z=c(0,1), t=c(0,3))
   X <- runifpointx(50, w)

[Package spatstat.random version 3.3-2 Index]