spatstat.random-internal {spatstat.random}	R Documentation

Internal spatstat.random functions

Description

Internal spatstat.random functions.

Usage

change.default.expand(x, newdefault)
datagen.runifpointOnLines(n, L)
datagen.runifpoisppOnLines(lambda, L)
datagen.rpoisppOnLines(lambda, L, lmax, ..., check)
default.clipwindow(object, epsilon)
expandwinPerfect(W, expand, amount)
detect.par.format(par, native, generic)
fakeNeyScot(Y, lambda, win, saveLambda, saveparents)
getRandomFieldsModelGen(model)
handle.rshift.args(W, ..., radius, width, height,
                   edge, clip, edgedefault)
HermiteCoefs(order)
is.cadlag(s)
is.expandable(x)
## S3 method for class 'rmhmodel'
is.expandable(x) 
kraever(package, fatal)
kraeverRandomFields()
MultiPair.checkmatrix(mat, n, matname, naok, zerook, asymmok)
optimalinflation(clusters, mod, rD)
## S3 method for class 'rmhcontrol'
print(x, ...)
## S3 method for class 'rmhexpand'
print(x, ..., prefix=TRUE)
## S3 method for class 'rmhmodel'
print(x, ...)
## S3 method for class 'rmhstart'
print(x, ...)
## S3 method for class 'rmhInfoList'
print(x, ...)
## S3 method for class 'summary.rmhexpand'
print(x, ...)
RandomFieldsSafe()
reheat(model, invtemp)
resolve.vargamma.shape(..., nu.ker, nu.pcf, nu, allow.nu, allow.default)
retrieve.param(desired, aliases, ..., par)
rMaternInhibition(type, kappa, r, win, stationary, ..., nsim, drop)
rMatClustHom(kappa, mu, R, W, ..., nsim, drop, inflate, saveparents)
rThomasHom(kappa, mu, sigma, W, ...,
           nsim, drop, inflate, saveparents, maxinflate)
rCauchyHom(kappa, mu, scale, W, ...,
           nsim, drop, inflate, saveparents, maxinflate)
rGRFcircembed(W, mu, var, corrfun, ..., nsim, drop)
rGRFgauss(W, mu, var, scale, ..., nsim, drop)
rGRFexpo(W, mu, var, scale, ..., nsim, drop)
rGRFstable(W, mu, var, scale, alpha, ..., nsim, drop)
rGRFgencauchy(W, mu, var, scale, alpha, beta, ..., nsim, drop)
rGRFmatern(W, mu, var, scale, nu, ..., nsim, drop)
rPoissonClusterEngine(kappa, expand, rcluster, win, ...,
                      nsim, drop, saveparents, kappamax, lmax, rmax) 
RmhExpandRule(nama)
## S3 method for class 'rmhcontrol'
rmhcontrol(...) 
## S3 method for class 'list'
rmhcontrol(...) 
rmhEngine(InfoList, ..., verbose, kitchensink, preponly, snoop,
                           overrideXstart, overrideclip) 
rmhResolveControl(control, model)
rmhResolveExpansion(win, control, imagelist, itype)
rmhResolveTypes(model, start, control)
rmhsnoop(..., Wsim, Wclip, R, xcoords, ycoords, mlevels, mcodes, irep, itype, 
     proptype, proplocn, propmark, propindx, numerator, denominator,
     panel.only)
rmhSnoopEnv(Xinit, Wclip, R)
## S3 method for class 'rmhmodel'
rmhmodel(model, ...) 
## S3 method for class 'rmhstart'
rmhstart(start, ...) 
## S3 method for class 'list'
rmhstart(start, ...) 
rmpoint.I.allim(n, f, types, nsim)
rpoint.multi(n, f, fmax, marks, win, giveup, verbose, warn, nsim, drop)
rthinEngine(X, P, ..., nsim, drop, Pmax, na.zero, what, fatal, warn)
runifpoispp(lambda, win, ..., nsim, drop)
runifpoisppOnLines(lambda, L, nsim, drop)
spatstatClusterModelInfo(name, onlyPCP)
spatstatClusterSimInfo(name)
spatstatClusterSimModelMatch(name, verbose)
spatstatRmhInfo(cifname)
summarise.trend(trend, w, a)
## S3 method for class 'rmhexpand'
summary(object, ...)
thinjump(n, p)
thinParents(X, P, Pmax)
## S3 method for class 'rmhstart'
update(object, ...)
validate.kappa.mu(kappa, mu, kappamax, mumax,
                  win, expand, ..., context)

Details

These internal spatstat.random functions should not be called directly by the user. Their names and capabilities may change without warning from one version of spatstat.random to the next.

Value

The return values of these functions are not documented, and may change without warning.

[Package spatstat.random version 3.3-2 Index]