update.rmhcontrol {spatstat.random}	R Documentation

Update Control Parameters of Metropolis-Hastings Algorithm

Description

update method for class "rmhcontrol".

Usage

  ## S3 method for class 'rmhcontrol'
update(object, ...)

Arguments

object
Object of class "rmhcontrol" containing control parameters for a Metropolis-Hastings algorithm.

...
Arguments to be updated in the new call to rmhcontrol.

Details

This is a method for the generic function update for the class "rmhcontrol". An object of class "rmhcontrol" describes a set of control parameters for the Metropolis-Hastings simulation algorithm. See rmhcontrol).

update.rmhcontrol will modify the parameters specified by object according to the new arguments given.

Value

Another object of class "rmhcontrol".

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

Examples

  a <- rmhcontrol(expand=1)
  update(a, expand=2)

[Package spatstat.random version 3.3-2 Index]