will.expand {spatstat.random}	R Documentation

Test Expansion Rule

Description

Determines whether an expansion rule will actually expand the window or not.

Usage

will.expand(x)

Arguments

x
Expansion rule. An object of class "rmhexpand".

Details

An object of class "rmhexpand" describes a rule for expanding a simulation window. See rmhexpand for details.

One possible expansion rule is to do nothing, i.e. not to expand the window.

This command inspects the expansion rule x and determines whether it will or will not actually expand the window. It returns TRUE if the window will be expanded.

Value

Logical value.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

See Also

rmhexpand, expand.owin

Examples

  x <- rmhexpand(distance=0.2)
  y <- rmhexpand(area=1)
  will.expand(x)
  will.expand(y)

[Package spatstat.random version 3.3-2 Index]