rpoispp3 {spatstat.random}	R Documentation

Generate Poisson Point Pattern in Three Dimensions

Description

Generate a random three-dimensional point pattern using the homogeneous Poisson process.

Usage

rpoispp3(lambda, domain = box3(), nsim=1, drop=TRUE)

Arguments

lambda
Intensity of the Poisson process. A single positive number.

domain
Three-dimensional box in which the process should be generated. An object of class "box3".

nsim
Number of simulated realisations to be generated.

drop
Logical. If nsim=1 and drop=TRUE (the default), the result will be a point pattern, rather than a list containing a point pattern.

Details

This function generates a realisation of the homogeneous Poisson process in three dimensions, with intensity lambda (points per unit volume).

The realisation is generated inside the three-dimensional region domain which currently must be a rectangular box (object of class "box3").

Value

If nsim = 1 and drop=TRUE, a point pattern in three dimensions (an object of class "pp3"). If nsim > 1, a list of such point patterns.

Note

The intensity lambda is the expected number of points per unit volume.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net.

See Also

runifpoint3, pp3, box3

Examples

   X <- rpoispp3(50)

[Package spatstat.random version 3.3-2 Index]