runifpoint3 {spatstat.random}	R Documentation

Generate N Uniform Random Points in Three Dimensions

Description

Generate a random point pattern containing n independent, uniform random points in three dimensions.

Usage

runifpoint3(n, domain = box3(), nsim=1, drop=TRUE)

Arguments

n
Number of points to be generated.

domain
Three-dimensional box in which the process should be generated. An object of class "box3".

nsim
Number of simulated realisations to be generated.

drop
Logical. If nsim=1 and drop=TRUE (the default), the result will be a point pattern, rather than a list containing a point pattern.

Details

This function generates n independent random points, uniformly distributed in the three-dimensional box domain.

Value

If nsim = 1 and drop=TRUE, a point pattern in three dimensions (an object of class "pp3"). If nsim > 1, a list of such point patterns.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net.

See Also

rpoispp3, pp3, box3

Examples

   X <- runifpoint3(50)

[Package spatstat.random version 3.3-2 Index]